Add files via upload
fixed some issues that showed up when starting the app
